requirements.txt
appdirs==1.4.3
asgiref==3.2.10
boto3==1.15.16
botocore==1.18.16
CacheControl==0.12.6
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1
django-environ==0.4.5
django-extensions==3.0.8
django-filter==2.3.0
django-import-export==2.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.2
html5lib==1.0.1
idna==2.8
importlib-metadata==2.0.0
ipaddr==2.2.0
iso3166==1.0.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
lockfile==0.12.2
mapbox==0.18.0
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==0.6.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
polyline==1.4.0
progress==1.5
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
Pyphen==0.9.5
python-dateutil==2.8.1
python-decouple==3.1
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
simplejson==3.17.2
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tablib==2.0.0
tinycss2==1.0.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.8
WeasyPrint==51
webencodings==0.5.1
zipp==3.3.0
drf-yasg==1.20.0
django-extensions==3.0.8
django-widget-tweaks==1.4.8
xhtml2pdf
fcm-django==0.3.4
pyfcm==1.4.7
